+++ apps/explorer/src/app/routes/route-names.tsx
@@ -1,0 +1,12 @@
+export const Routes = {
+  HOME: '/',
+  TX: 'txs',
+  BLOCKS: 'blocks',
+  PARTIES: 'parties',
+  VALIDATORS: 'validators',
+  ASSETS: 'assets',
+  GENESIS: 'genesis',
+  GOVERNANCE: 'governance',
+  MARKETS: 'markets',
+  NETWORK_PARAMETERS: 'network-parameters',
+};
